Added typecast to f_getpid call.
git-svn-id: http://svn.code.sf.net/p/flaim/code/trunk@545 0109f412-320b-0410-ab79-c3e0c5ffbbe6

+++ flaim/src/fsysdata.cpp
@@ -596,7 +596,7 @@ FLMEXP RCODE FLMAPI FlmStartup( void)
 	}
 	
 #ifdef FLM_USE_NICI
-	iHandle  = f_getpid();
+	iHandle  = (int)f_getpid();
 
 	// Initialize NICI
 	if (CCS_Init(&iHandle))
